+++ package.json
@@ -32,10 +32,6 @@
     "sass": "^1.50.1",
     "typescript": "~4.1.5"
   },
-  "resolutions": {
-    "@types/react": "17.0.2",
-    "@types/react-dom": "17.0.2"
-  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
