+++ src/custom.d.ts
@@ -1,0 +1,8 @@
+export {};
+
+declare global {
+  // eslint-disable-next-line no-unused-vars
+  interface Window {
+    msw: Record<string, any>;
+  }
+}
