+++ src/modules/auth/__tests__/Login.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { render, screen, waitFor } from '../../../testUtils/utils';
+import { render, screen, waitFor } from '../../../__mocks__/utils';
 import Login from '../Login';
-import server from '../../../testUtils/server';
-import { userWithTokens } from '../../../testUtils/data';
+import server from '../../../__mocks__/server';
+import { userWithTokens } from '../../../__mocks__/data';
 import { Routing } from '../../common/routing';
 import routes from '../../common/routing/routes';
