+++ src/modules/dashboard/pages/home/Dashboard.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen } from '../../../../testUtils/utils';
+import { render, screen } from '../../../../__mocks__/utils';
 import DashboardHome from './DashboardHome';
 
 describe('Dashboard Overview', () => {
