+++ src/modules/dashboard/pages/profile/Profile.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { render, screen, waitFor } from '../../../../testUtils/utils';
+import { render, screen, waitFor } from '../../../../__mocks__/utils';
 import Profile from './Profile';
-import server from '../../../../testUtils/server';
+import server from '../../../../__mocks__/server';
 import { Routing } from '../../../common/routing';
 import routes from '../../../common/routing/routes';
